fix: share button
