+++ apps/website/components/templates/gate-view/gate-view.tsx
@@ -4,7 +4,6 @@ import { useState, useEffect } from 'react';
 
 import { PartialDeep } from 'type-fest';
 
-import ShareIcon from '@mui/icons-material/IosShare';
 import ViewInArIcon from '@mui/icons-material/ViewInAr';
 import {
   Avatar,
@@ -16,16 +15,14 @@ import {
   Box,
   Divider,
   Button,
-  Snackbar,
-  IconButton,
 } from '@mui/material';
 
 import { useAuth } from '../../../../website/providers/auth';
 import { useMint } from '../../../hooks/use-mint';
-import { useSnackbar } from '../../../hooks/use-snackbar';
 import { Gates } from '../../../services/graphql/types.generated';
 import { AvatarFile } from '../../atoms/avatar-file';
 import CircularProgressWithLabel from '../../atoms/circular-progress-label';
+import { ShareButton } from '../../atoms/share-button';
 import GateCompletedModal from '../../organisms/gates/view/modals/gate-completed';
 import { Task, TaskGroup } from './tasks';
 
@@ -36,7 +33,6 @@ type Props = {
 export function GateViewTemplate({ gate }: Props) {
   const { t } = useTranslation();
   const taskIds = gate.tasks.map((task) => task.id);
-  const snackbar = useSnackbar();
   const { me } = useAuth();
   const { mint } = useMint();
 
@@ -65,22 +61,6 @@ export function GateViewTemplate({ gate }: Props) {
       handleOpen();
     }
   }, [taskIds, me?.task_progresses, completedTasksCount]);
-
-  const onShare = () => {
-    const data = {
-      title: `${gate.title} @ Gateway`,
-      url: window.location.href,
-    };
-    try {
-      if (navigator?.share && navigator.canShare(data)) {
-        navigator.share(data);
-      } else if (navigator?.clipboard && navigator.clipboard) {
-        snackbar.onOpen({ message: 'Copied link!' });
-      }
-    } catch (e) {
-      console.error(e);
-    }
-  };
 
   return (
     <Grid container height="100%" sx={{ flexWrap: 'nowrap' }}>
@@ -140,17 +120,7 @@ export function GateViewTemplate({ gate }: Props) {
               ))}
             </Box>
             <Stack>
-              <IconButton
-                sx={{
-                  p: 0,
-                }}
-                onClick={onShare}
-                key="share"
-              >
-                <Avatar>
-                  <ShareIcon sx={{ mt: -0.25 }} />
-                </Avatar>
-              </IconButton>
+              <ShareButton title={`${gate.title} @ Gateway`} />
             </Stack>
           </Stack>
         </Box>
@@ -298,15 +268,6 @@ export function GateViewTemplate({ gate }: Props) {
           ))}
         </TaskGroup>
       </Grid>
-      <Snackbar
-        anchorOrigin={{
-          vertical: snackbar.vertical,
-          horizontal: snackbar.horizontal,
-        }}
-        open={snackbar.open}
-        onClose={snackbar.handleClose}
-        message={snackbar.message}
-      />
     </Grid>
   );
 }
